dd: only write when changes are needed
Useful for hanlding ZFS ZVOLs. Writing the same data over causes
unnecessary disk space consumption. First read what is on the target
device and only write the new data when it is different.

This is useful expecially for writing to clones of snapshots when the
new data is mostly the same.